requirements.txt
Django==4.2.7
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
django-filter==23.2
PyYAML==6.0.1
ujson==5.8.0


